requirements.txt
PyYAML~=6.0.1
streamlit~=1.26.0
yfinance==0.2.31
alpha_vantage==2.3.1
plotly==5.17.0
streamlit-authenticator==0.2.3
matplotlib==3.8.0
pandas==2.1.1
googletrans==3.1.0-alpha
newspaper3k==0.2.8
nltk==3.7
requests==2.31.0
matplotlib==3.8.0
ipywidgets==8.1.1
googletrans==4.0.0rc1
nltk==3.8.1
protobuf==3.20.*
transformers==4.33.1
extra-streamlit-components==0.1.60
torchaudio==2.1.0
torchvision==0.16.0
st-pages==0.4.5
